Move common helper into util.py
